requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bleach==6.1.0
blinker==1.7.0
bokeh==3.3.4
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
fonttools==4.50.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
grpcio==1.53.0
grpcio-tools==1.53.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
linkify-it-py==2.0.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdit-py-plugins==0.4.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
openai==1.13.3
packaging==23.2
pandas==2.2.1
panel==1.3.8
param==2.0.2
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==4.21.12
pyarrow==15.0.1
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyvista==0.43.4
pyviz_comms==3.0.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scooby==0.9.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stability-sdk==0.8.5
stpyvista==0.0.15
streamlit==1.32.1
streamlit-float==0.3.2
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.12
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
triton==2.2.0
types-python-dateutil==2.8.19.20240311
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
urllib3==2.2.1
validators==0.22.0
vtk==9.3.0
wasmtime==12.0.0
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.16.0
xyzservices==2023.10.1
zipp==3.17.0
